* Updated some javadoc * fixed fields of BrickLinkPart to match TSV field titles and also added @JsonPropertyOrder to add the order expected of the fields in the TSV data * Dimensions in Part object should be floats not integers * Created base parser that all TSV parsing for BrickLink data will use and added parsers for Part and Category * Fixed partFactory to extract the correct dimensions from the dimensions string within BrickLinkPart * Added some unit tests * Added dependency on jackson-dataformat-csv for parsers module * Added dependency on jackson-annotations for Model module

+++ brickorganiser-model/src/main/java/com/github/codersparks/brickorganiser/model/Part.java
@@ -1,6 +1,5 @@
 package com.github.codersparks.brickorganiser.model;
 
-import com.github.codersparks.brickorganiser.model.bricklink.BrickLinkItem;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -30,8 +29,8 @@ public class Part {
     @Indexed
     private String category;
 
-    private Integer dimX = null;
-    private Integer dimY = null;
-    private Integer dimZ = null;
+    private Float dimX = null;
+    private Float dimY = null;
+    private Float dimZ = null;
 
 }
